requirements.txt
requests==2.25.0
waitress==1.4.4
xlrd==1.2.0
xlwings==0.21.4
imutils==0.5.3
Flask==1.1.2
numpy==1.19.2
matplotlib==3.3.3
~pencv_python==4.4.0.46
XlsxWriter==1.3.7
pywin32==300
Flask_Cors==3.0.9
pandas==1.1.5
pyodbc==4.0.30
Pillow==8.1.0
secrets==1.0.2
